+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/auth/domain/AuthService.java
@@ -5,6 +5,8 @@ import org.haedal.zzansuni.global.jwt.JwtToken;
 import org.haedal.zzansuni.global.jwt.JwtUser;
 import org.haedal.zzansuni.global.jwt.JwtUtils;
 import org.haedal.zzansuni.user.domain.*;
+import org.haedal.zzansuni.user.domain.port.UserReader;
+import org.haedal.zzansuni.user.domain.port.UserStore;
 import org.springframework.data.util.Pair;
 import org.springframework.lang.NonNull;
 import org.springframework.lang.Nullable;
