--- zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/user/domain/UserStore.java
@@ -1,5 +1,0 @@
-package org.haedal.zzansuni.user.domain;
-
-public interface UserStore {
-    User store(User user);
-}
