+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/user/domain/port/UserStore.java
@@ -1,0 +1,7 @@
+package org.haedal.zzansuni.user.domain.port;
+
+import org.haedal.zzansuni.user.domain.User;
+
+public interface UserStore {
+    User store(User user);
+}
