+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/user/infrastructure/UserStoreImpl.java
@@ -2,7 +2,7 @@ package org.haedal.zzansuni.user.infrastructure;
 
 import lombok.RequiredArgsConstructor;
 import org.haedal.zzansuni.user.domain.User;
-import org.haedal.zzansuni.user.domain.UserStore;
+import org.haedal.zzansuni.user.domain.port.UserStore;
 import org.springframework.stereotype.Component;
 
 @Component
